requirements.txt
gradio
pillow
